Renamed the Extract Text examples to Batch Read File
Computer Vision - C#

The naming better reflects the API names. Explanations were added as to the differences.
The variables for these examples were left alone, since the variable names are better if they match the snippet tag names.
But for the function definition and call in main(), know that any snippet tag with Extract Text name is really Batch Read File.